chore[#1]: 배포 환경 설정
